Rename location list to main activity
