+++ app/src/main/AndroidManifest.xml
@@ -28,14 +28,6 @@
             <!-- <category android:name="android.intent.category.LAUNCHER" /> -->
             <!-- </intent-filter> -->
         </activity>
-        <activity
-            android:name=".controller.MainActivity"
-            android:label="@string/title_activity_main"
-            android:theme="@style/AppTheme.NoActionBar">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="kantwonskids.donationtrackerg14b.controller.LoginActivity" />
-        </activity>
         <activity android:name=".controller.WelcomeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -45,17 +37,17 @@
         </activity>
         <activity android:name=".controller.RegistrationActivity" />
         <activity
-            android:name=".controller.LocationListActivity"
+            android:name=".controller.MainActivity"
             android:label="@string/title_location_list"
             android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity
             android:name=".controller.LocationDetailActivity"
             android:label="@string/title_location_detail"
-            android:parentActivityName=".controller.LocationListActivity"
+            android:parentActivityName=".controller.MainActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="kantwonskids.donationtrackerg14b.controller.LocationListActivity" />
+                android:value="kantwonskids.donationtrackerg14b.controller.MainActivity" />
         </activity>
         <activity
                 android:name=".controller.InventoryActivity" >
