+++ app/src/main/java/kantwonskids/donationtrackerg14b/controller/LoginActivity.java
@@ -141,7 +141,7 @@ public class LoginActivity extends AppCompatActivity {
      */
     private void login(User potentialUser) {
 //        Intent intent = new Intent(this, MainActivity.class);
-        Intent intent = new Intent(this, LocationListActivity.class);
+        Intent intent = new Intent(this, MainActivity.class);
         intent.putExtra("CURRENT_USER", (Parcelable)potentialUser);
         startActivity(intent);
     }
