+++ app/src/main/res/layout-w900dp/location_list.xml
@@ -10,7 +10,7 @@
     android:divider="?android:attr/dividerHorizontal"
     android:orientation="horizontal"
     android:showDividers="middle"
-    tools:context=".controller.LocationListActivity">
+    tools:context=".controller.MainActivity">
 
     <!--
     This layout is a two-pane layout for the Locations
@@ -27,7 +27,7 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context="kantwonskids.donationtrackerg14b.controller.LocationListActivity"
+        tools:context="kantwonskids.donationtrackerg14b.controller.MainActivity"
         tools:listitem="@layout/location_list_content" />
 
     <FrameLayout
